requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.8

# Web Framework
fastapi==0.109.2
uvicorn==0.27.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Authentication
PyJWT==2.10.1

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0

# Utilities
aiohttp==3.9.1
requests==2.31.0

# Development (optional)
black==23.12.1
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.23.2

# Production
gunicorn==21.2.0

# Logging & Monitoring
python-json-logger==2.0.7
